refactor: ColorPicker로 네이밍 변경
