+++ src/components/organism/CardEditModal/HashTagsInputs.jsx
@@ -2,33 +2,6 @@ import React, { useState } from 'react';
 import styled from '@emotion/styled';
 import Common from '@styles';
 
-const InputContainer = styled.div`
-  display: grid;
-  justify-items: center;
-  align-items: center;
-  grid-gap: 8px;
-  grid-template-columns: repeat(3, 1fr);
-  @media ${Common.media.sm} {
-    grid-template-columns: repeat(2, 1fr);
-  }
-`;
-
-const Input = styled.input`
-  display: block;
-  padding: 0 8px;
-  width: 114px;
-  height: 40px;
-  font-size: 14px;
-  color: ${({ color }) => color};
-  border-radius: 4px;
-  border: 2px solid ${({ color }) => color};
-  background-color: transparent;
-  box-sizing: border-box;
-  @media ${Common.media.sm} {
-    width: 140px;
-  }
-`;
-
 const HASHTAG = {
   0: 'first',
   1: 'second',
@@ -93,5 +66,32 @@ const HashTagsInputs = React.memo(
     );
   },
 );
+
+const InputContainer = styled.div`
+  display: grid;
+  justify-items: center;
+  align-items: center;
+  grid-gap: 8px;
+  grid-template-columns: repeat(3, 1fr);
+  @media ${Common.media.sm} {
+    grid-template-columns: repeat(2, 1fr);
+  }
+`;
+
+const Input = styled.input`
+  display: block;
+  padding: 0 8px;
+  width: 114px;
+  height: 40px;
+  font-size: 14px;
+  color: ${({ color }) => color};
+  border-radius: 4px;
+  border: 2px solid ${({ color }) => color};
+  background-color: transparent;
+  box-sizing: border-box;
+  @media ${Common.media.sm} {
+    width: 140px;
+  }
+`;
 
 export default HashTagsInputs;
